enums refactor and wifi fucking up fix
