--- Core/include/enumAngleSource.h
@@ -1,9 +1,0 @@
-#ifndef ENUM_ANGLE_SOURCE_H
-#define ENUM_ANGLE_SOURCE_H
-enum class EnumAngleSource
-{
-    IMU,
-    ENCODER,
-};
-
-#endif
