--- Core/include/enumIMU.h
@@ -1,12 +1,0 @@
-#ifndef ENUM_IMU_H
-#define ENUM_IMU_H
-enum class enumIMU
-{
-    HIP_L    =   0,
-    KNEE_L   =   1,
-    HIP_R    =   2,
-    KNEE_R   =   3,
-    EXO_BACK =   4    
-};
-
-#endif
