--- Core/include/enumSonarState.h
@@ -1,9 +1,0 @@
-#ifndef ENUM_SONAR_STATE_H
-#define ENUM_SONAR_STATE_H
-enum class EnumSonarState
-{
-    SQUAT,
-    WALK,
-};
-
-#endif
