+++ Core/include/enums.h
@@ -1,5 +1,26 @@
-#ifndef ENUM_SCREEN_CALLBACKS_H
-#define ENUM_SCREEN_CALLBACKS_H
+#ifndef ENUMS_H
+#define ENUMS_H
+enum class EnumAngleSource
+{
+    IMU,
+    ENCODER,
+};
+
+enum class enumIMU
+{
+    HIP_L    =   0,
+    KNEE_L   =   1,
+    HIP_R    =   2,
+    KNEE_R   =   3,
+    EXO_BACK =   4    
+};
+
+enum class EnumSonarState
+{
+    SQUAT,
+    WALK,
+};
+
 enum class EnumScreenCallback
 {
     BUTTON_TOGGLE_MOTOR,
