+++ Core/include/touchScreen.h
@@ -6,7 +6,7 @@
 #include <string>
 
 
-#include "enumScreenCallbacks.h"
+#include "enums.h"
 
 using namespace std;
