+++ Core/platformio.ini
@@ -13,6 +13,7 @@ platform = espressif32
 board = esp32doit-devkit-v1
 framework = arduino
 lib_deps = 
+	Wifi
 	SPI
 	arduino-libraries/Arduino_JSON@^0.1.0
 	adafruit/Adafruit Unified Sensor@^1.1.9
